Fix signature to allow us switch between stored_function_result representations

+++ server/migrations/20180527_183716_add_stored_function_result.sql
@@ -1,9 +1,9 @@
 CREATE TABLE IF NOT EXISTS
 function_results
 ( canvas_id UUID REFERENCES canvases(id) NOT NULL
-, tlid INT NOT NULL
+, tlid BIGINT NOT NULL
 , fnname TEXT NOT NULL
-, id INT NOT NULL
+, id BIGINT NOT NULL
 , hash TEXT NOT NULL
 , timestamp TIMESTAMPTZ NOT NULL
 , value TEXT NOT NULL
